Clean up dependencies of main crate
