+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+  "cargo-creusot",
   "creusot",
   "creusot-contracts",
   "creusot-contracts-proc",
