+++ cargo-creusot/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "cargo-creusot"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde_json = { version = "1.0" }
+clap = { version = "4.2", features = ["derive", "env"] }
+creusot = { path = "../creusot", version = "*" }
+toml = "0.5.8"
+env_logger = "*"
+serde = { version = "1.0", features = ["derive"] }
