+++ creusot-contracts/src/ghost_ptr.rs
@@ -69,8 +69,8 @@ impl<T: ?Sized> GhostPtrToken<T> {
 
     /// Shrinks the view of the `self` so that it's model is now new-model
     #[trusted]
-    #[requires(new_model.subset(self@))]
-    #[ensures(result@ == *new_model)]
+    #[requires(_new_model.subset(self@))]
+    #[ensures(result@ == *_new_model)]
     pub fn shrink_token_ref(&self, _new_model: Ghost<FMap<*const T, T>>) -> &GhostPtrToken<T> {
         self
     }
@@ -109,10 +109,10 @@ impl<T: ?Sized> GhostPtrToken<T> {
     }
 
     #[trusted]
-    #[ensures((*self)@.disjoint(other@))]
+    #[ensures((*self)@.disjoint(_other@))]
     // Since we had full permission to and all of the entries in `self` and `other` simultaneously,
     // no pointer could have been in both
-    #[ensures((^self)@ == (*self)@.union(other@))]
+    #[ensures((^self)@ == (*self)@.union(_other@))]
     pub fn merge(&mut self, _other: GhostPtrToken<T>) {}
 
     /// Leaks memory iff the precondition fails
