+++ creusot/Cargo.toml
@@ -6,16 +6,14 @@ edition = "2021"
 
 [dependencies]
 itertools = "*"
-env_logger = "*"
 log = "0.4"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0" }
 heck = "0.4"
 petgraph = "0.6"
 indexmap = { version = "1.7.0", features = ["serde"] }
 toml = "0.5.8"
 why3 = { path = "../why3", features = ["serialize"] }
-clap = { version = "3.2", features = ["derive", "env"] }
+clap = { version = "4.2", features = ["derive", "env"] }
 creusot-metadata = { path = "../creusot-metadata" }
 lazy_static = "1.4.0"
