+++ why3tests/Cargo.toml
@@ -9,7 +9,7 @@ glob = "*"
 assert_cmd = "1.0"
 termcolor = "1.1"
 git2 = "0.14.4"
-clap = { version = "3.2.8", features = ["env", "derive"]}
+clap = { version = "4.2", features = ["env", "derive"]}
 
 [[test]]
 test = false
